examples/viewer: Fix not being able to playback more than 200 frames
Seems that when file size is large, several variables of type int will
overflow. This happens when recording 'lr-native' (1024x1024) and more
than 200 frames (205) to be precise.
Use size_t which should contain 8 bytes of unsinged integer. This should
handle situations with larger files.
Files can't be infinite large since viewer has a upper limit of 300
frames to record.

+++ examples/tof-viewer/include/ADIToFRecorder.h
@@ -173,10 +173,10 @@ class ADIToFRecorder {
     bool m_saveBinaryFormat = false;
     bool m_finishRecording = true;
     bool _stopPlayback = false;
-    int currentPBPos = 0;
+    size_t currentPBPos = 0;
     int totalBits = 0;
     int m_numberOfFrames;
-    int fileSize = 0;
+    size_t fileSize = 0;
     uint16_t m_sizeOfHeader;
 
   private:
